revert caom2.Plane.polarization_dimension type to integer in database…

+++ caom2persistence/src/main/resources/postgresql/caom2.Plane.sql
@@ -72,7 +72,8 @@ create table caom2.Plane
 
 -- polarization
     polarization_states     varchar(34),
-    polarization_dimension  bigint,
+-- this is long in the code but integer in db for historical reasons
+    polarization_dimension  integer,
 
 -- optimisation
     dataReadAccessGroups tsvector default '',
